Add a SoloX.GeneratorTools.Core.Test.Helpers project.
