+++ src/tests/SoloX.GeneratorTools.Core.CSharp.ITest/SoloX.GeneratorTools.Core.CSharp.ITest.csproj
@@ -27,6 +27,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\libs\SoloX.GeneratorTools.Core.CSharp\SoloX.GeneratorTools.Core.CSharp.csproj" />
     <ProjectReference Include="..\SoloX.GeneratorTools.Core.CSharp.UTest\SoloX.GeneratorTools.Core.CSharp.UTest.csproj" />
+    <ProjectReference Include="..\SoloX.GeneratorTools.Core.Test.Helpers\SoloX.GeneratorTools.Core.Test.Helpers.csproj" />
   </ItemGroup>
 
   <ItemGroup>
